vps-agent/requirements.txt
anthropic>=0.40.0
psycopg2-binary>=2.9.9
aiohttp>=3.10.0
feedparser>=6.0.11
numpy>=1.26.0
python-dotenv>=1.0.1
